@@ -1,18 +1,20 @@
 {
-    "name": "@slimio/pretty-json",
+    "name": "@topcli/pretty-json",
     "version": "1.3.0",
     "description": "Stdout JSON in your terminal",
-    "main": "index.js",
-    "scripts": {
+    "main": "./index.js",
+    "type": "module",
+     "scripts": {
         "start": "node index.js",
+        "lint": "eslint .",
         "prepublishOnly": "pkg-ok",
-        "test": "cross-env psp && jest --coverage",
+        "test": "node --test test/test.js",
         "coverage": "codecov",
         "doc": "jsdoc -c ./jsdoc.json -r -R ./README.md -P ./package.json --verbose"
     },
     "repository": {
         "type": "git",
-        "url": "git+https://github.com/SlimIO/Pretty-JSON.git"
+        "url": "git+https://github.com/TopCli/Pretty-JSON.git"
     },
     "keywords": [
         "pretty",
@@ -27,37 +29,24 @@
         "index.d.ts",
         "src/utils.js"
     ],
-    "author": "SlimIO",
+    "author": "GENTILHOMME Thomas <gentilhomme.thomas@gmail.com>",
     "license": "MIT",
     "bugs": {
-        "url": "https://github.com/SlimIO/Pretty-JSON/issues"
+        "url": "https://github.com/TopCli/Pretty-JSON/issues"
     },
-    "homepage": "https://github.com/SlimIO/Pretty-JSON#readme",
+    "homepage": "https://github.com/TopCli/Pretty-JSON#readme",
     "dependencies": {
-        "@slimio/is": "^1.5.1",
-        "kleur": "^3.0.3"
+        "@slimio/is": "^2.0.0",
+        "kleur": "^4.1.5"
     },
     "devDependencies": {
-        "@commitlint/cli": "^8.3.5",
-        "@commitlint/config-conventional": "^8.3.4",
         "@escommunity/minami": "^1.0.0",
-        "@slimio/eslint-config": "^4.1.0",
-        "@slimio/psp": "^0.11.1",
+        "@nodesecure/eslint-config": "^1.7.0",
         "codecov": "^3.7.0",
-        "cross-env": "^7.0.2",
-        "eslint": "^7.0.0",
-        "husky": "^4.2.5",
-        "jest": "^26.0.1",
-        "jsdoc": "^3.6.4",
-        "pkg-ok": "^2.3.1"
+        "eslint": "^8.32.0",
+        "jsdoc": "^4.0.2"
     },
     "engines": {
-        "node": ">=12"
-    },
-    "husky": {
-        "hooks": {
-            "pre-push": "cross-env eslint index.js && npm test",
-            "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
-        }
+        "node": ">=16"
     }
 }
